Add live watch and debug-helper
devtool: source-map, this creates amain.js.map file in the dist folder, and it makes sure that all the errors and/or console.logs from the origin files will be displayed correctly in the console with the correct pathname and correct line. Conclusion: Important step for debugging!
Add --watch after the build: webpack in the package.json file.
This is to make your bundle-file main.js updated automatically whenever a change is made in one of the original files that are bundled. this replaces the 'npm run build' in the console everytime after you made a change in the files